+++ УчетПокупокУслугКлиентов.cs
@@ -178,7 +178,7 @@ namespace Client
 
             sql = "SELECT НС, ФИО FROM Портье WHERE НС=@НС";
             ComboBoxDataForFill Портье = new ComboBoxDataForFill(sql, "ФИО", "НС");
-            Портье.paramsForSQLQuery.Add(new MySqlParameter("@НС", MySqlDbType.Int32) { Value = Convert.ToInt16(lbWhoLogged.Text) });
+            Портье.paramsForSQLQuery.Add(new MySqlParameter("@НС", MySqlDbType.Int32) { Value = Convert.ToInt32(lbWhoLogged.Text) });
             LoadCombo(Портье, кбНС);
 
 
@@ -261,7 +261,7 @@ namespace Client
                 // Загружаем портье
                 sql = "SELECT НС, ФИО FROM Портье WHERE НС=@НС";
                 ComboBoxDataForFill Портье = new ComboBoxDataForFill(sql, "ФИО", "НС");
-                Портье.paramsForSQLQuery.Add(new MySqlParameter("@НС", MySqlDbType.Int32) { Value = Convert.ToInt16(lbWhoLogged.Text) });
+                Портье.paramsForSQLQuery.Add(new MySqlParameter("@НС", MySqlDbType.Int32) { Value = Convert.ToInt32(lbWhoLogged.Text) });
                 LoadCombo(Портье, кбНС);
 
                 await LoadServicesByRequest(НЗаявки.Value);
